Ammo and health values change and slide

+++ Assets/Scripts/Abstracts/BallisticWeapon.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 public abstract class BallisticWeapon : Weapon
 {
-    public int MaxClipSize { get; set; }
-    public int CurrentClipSize { get; set; }
+    public float MaxClipSize { get; set; }
+    public float CurrentClipSize { get; set; }
     public float FireRate { get; set; }
 }
